Extract resource name constants and improve code organization
- Create resources.go with constants for resource names
- Move inlineAppConfigFile from import.go to yaml_utils.go
- Replace string literals with resource name constants
- Update doc comments for consistency and conciseness
- Use lowercase in warning messages for consistency
